+++ ZConfig/tests/test_schema.py
@@ -1130,25 +1130,33 @@ class SchemaTestCase(TestHelper, unittest.TestCase):
             </schema>
             """)
 
-    def test_section_example_is_error(self):
-        self.assertRaises(ZConfig.SchemaError, self.load_schema_text, """\
+    def test_section_example(self):
+        schema = self.load_schema_text("""\
             <schema>
-                <sectiontype name="abc"/>
+                <sectiontype name="abc">
+                    <example>  This is a sectiontype example  </example>
+                </sectiontype>
                 <section type="abc" name="def">
                     <example>  This is an example  </example>
                 </section>
             </schema>
             """)
+        self.assertEqual(schema.getinfo('def').sectiontype.example, 'This is a sectiontype example')
+        self.assertEqual(schema.getinfo('def').example, 'This is an example')
 
-    def test_multisection_example_is_error(self):
-        self.assertRaises(ZConfig.SchemaError, self.load_schema_text, """\
+    def test_multisection_example(self):
+        schema = self.load_schema_text("""\
             <schema>
-                <sectiontype name="abc"/>
+                <sectiontype name="abc">
+                    <example>  This is a sectiontype example  </example>
+                </sectiontype>
                 <multisection type="abc" name="*" attribute="things">
                     <example>  This is an example  </example>
                 </multisection>
             </schema>
             """)
+        self.assertEqual(schema[0][1].sectiontype.example, 'This is a sectiontype example')
+        self.assertEqual(schema[0][1].example, 'This is an example')
 
     def checkErrorText(self, schema, error_text):
         self.assertRaisesRegexp(ZConfig.SchemaError, error_text,