+++ CHANGES.rst
@@ -5,7 +5,9 @@ Change History for ZConfig
 3.2.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add support for ``example`` in ``section`` and ``multisection``, and
+  include those examples in generated documentation. See
+  https://github.com/zopefoundation/ZConfig/pull/5.
 
 
 3.2.0 (2017-06-22)